% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_editions.R
\name{getData_editions}
\alias{getData_editions}
\title{Harvest CBLOL editions data from Leaguepedia}
\usage{
getData_editions(Year = c(2014:2020))
}
\arguments{
\item{Year}{numeric. The year of the edition you want to access data}
}
\value{
A tibble containing: year, tournament (split/playoff), prize pool (R$), winner, runner up and league.
}
\description{
Creates a tibble containing Leaguepedia data on each CBLOL edition.
}
\examples{
a <- getData_editions()
b <- getData_editions(ano = c(2020,2018))
}
