% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_playersChampion.R
\name{getData_playersChampion}
\alias{getData_playersChampion}
\title{Harvest from Leaguepedia data of players per Champion played in games of CBLOL}
\usage{
getData_playersChampion(Role, Year, Split, Playerid = NULL, Champion = NULL)
}
\arguments{
\item{Role}{(character) The lane of the player. It should contain at least one of the five roles: "Top", "Jungle", "Mid", "Bot" and "Support".}

\item{Year}{(numeric) The year you want to access data(2015:2020).}

\item{Split}{(character) The split you want to access data: "Split 1", "Split 2", "Split 1 Playoffs" or "Split 2 Playoffs".}

\item{Playerid}{(character) The player you want to access data. By default it returns data on every player. Its very case sensitive and the playerid(s) should be passed exactly as in Leaguepedia}

\item{Champion}{(character) The champion you want to access data. By default it returns data on every champion. Its very case sensitive and the champion(s) name(s) should be passed exactly as in Leaguepedia}
}
\value{
A tibble containing: player, role, year, split, champion, number of games played with the champion, victories, defeats, win rate, kills, deaths, assists, KDA, CS per game, CS per minute, gold per game, gold per minute, kill participation, percentage of kills/team, percentage of gold/team and league.
}
\description{
Creates a tibble containing Leaguepedia data of players  per Champion played in games of CBLOL
}
\examples{
players_champion <- getData_playersChampion(Role = "Jungle", Year = 2020, Split = c("Split 2","Split 2 Playoffs"))
}
