% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_rosters.R
\name{getData_rosters}
\alias{getData_rosters}
\title{Harvest data of CBLOL rosters from Leaguepedia}
\usage{
getData_rosters(Playerid = NULL, Team = NULL, Split, Year, Role)
}
\arguments{
\item{Playerid}{(character)  Its very case sensitive and the playerid(s) have to be passed exactly as it is written at Leaguepedia}

\item{Team}{(character)  Its very case sensitive and the name(s) have to be passed exactly as it is written at Leaguepedia}

\item{Split}{(character)  Use if you want to specify a Split: "Split 1" or "Split 2".}

\item{Year}{(numeric) The year of the edition you want to access data (2015:2020)}

\item{Role}{(character)  Use if you want to specify one or more roles: "Top Laner", "Jungler", "Mid Laner", "Bot Laner", "Support" or "Coach".}
}
\value{
A tibble containing: country, playerid (id), name, team, year, split, role and league. If there is no data for the entry you specified it will return a message saying "There is no data for this entry".
}
\description{
Creates a tibble containing Leaguepedia data on CBLOL rosters on each edition
}
\examples{
rosters <- getData_rosters(Year = c(2015:2020),Team = "INTZ",Role = "Support", Split = c("Split 1","Split 2"))
}
