% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_titlesOrg.R
\name{getData_titlesOrg}
\alias{getData_titlesOrg}
\title{Harvest data of CBLOL wins per Organization from Leaguepedia}
\usage{
getData_titlesOrg(Team = NULL)
}
\arguments{
\item{Team}{character. Its very sensitive and the names have to be passed exactly as it is written at https://lol.gamepedia.com/Circuit_Brazilian_League_of_Legends}
}
\value{
A tibble containing: team, x1st_place, x2nd_place, x3rd_4th_place, league
}
\description{
Creates a tibble containing Leaguepedia data CBLOL wins per Organization
}
\examples{
a <- getData_titlesOrg()
b <- getData_titlesOrg(Team = "INTZ")
c <- getData_titlesOrg(Team = c("INTZ","paiN Gaming"))
}
