% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_titlesPlayer.R
\name{getData_titlesPlayer}
\alias{getData_titlesPlayer}
\title{Harvest data of CBLOL wins per player from Leaguepedia}
\usage{
getData_titlesPlayer(playerid = NULL)
}
\arguments{
\item{playerid}{character. Its very sensitive and the playerid(s) have to be passed exactly as it is written at https://lol.gamepedia.com/Circuit_Brazilian_League_of_Legends}
}
\value{
A tibble containing: player, titles, titles_as_player, titles_as_coach and league.
}
\description{
Creates a tibble containing Leaguepedia data on CBLOL wins per player
}
\examples{
a <- getData_titlesPlayer()
b <- getData_titlesPlayer(playerid = "brTT")
c <- getData_titlesPlayer(playerid = c("Kami","brTT"))
}
