% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_champion.R
\name{getData_champion}
\alias{getData_champion}
\title{Harvest data of champions in games of CBLOL from Leaguepedia}
\usage{
getData_champion(
  Role = c("Top", "Jungle", "Mid", "AD Carry", "Support"),
  Year = 2015:2020,
  Split = c("Split_1", "Split_1_Playoffs", "Split_2", "Split_2_Playoffs"),
  Champion = NULL
)
}
\arguments{
\item{Role}{(character) The lane where the champion was played. It should contain at least one of the five roles: "Top", "Jungle", "Mid", "AD Carry" and "Support".}

\item{Year}{(numeric) The year you want to access data.}

\item{Split}{(character) The split you want to access data: "Split_1", "Split_2", "Split_1_Playoffs" or "Split_2_Playoffs".}

\item{Champion}{(character) The champion you want to access data. By default it returns data on every champion. Its very case sensitive.}
}
\value{
A tibble containing: champion, number of games it was played, victories, defeats, win rate, kills, deaths, assists, KDA, CS per game, CS per minute, gold per game, gold per minute, kill participation, percentage of kills/team, percentage of gold/team, lane, year, split and league.
}
\description{
Creates a tibble containing Leaguepedia data on champions played in CBLOL games
}
\examples{
a <- getData_champion()
b <- getData_champion(
    Role = "Top",
    Year = 2020,
    Split = c("Split_1","Split_1_Playoffs")
)
c <- getData_champion(
Role = "AD Carry",
Year = 2019,
Champion = "Ashe"
)
}
