% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_editions.R
\name{getData_editions}
\alias{getData_editions}
\title{Harvest CBLOL editions data from Leaguepedia}
\usage{
getData_editions(
  Split = c("Split 1", "Split 2", "Split 1 Playoffs", "Split 2 Playoffs"),
  Year = c(2014:2020)
)
}
\arguments{
\item{Split}{(character) The split you want to access data: "Split 1", "Split 2", "Split 1 Playoffs" or "Split 2 Playoffs".}

\item{Year}{(numeric) The year you want to access data}
}
\value{
A tibble containing: year, tournament (split/playoff), prize pool (R$), winner, runner up and league.
}
\description{
Creates a tibble containing Leaguepedia data on each CBLOL edition.
}
\examples{
a <- getData_editions()
b <- getData_editions(ano = c(2020,2018))
}
