% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_editions.R
\name{getData_editions}
\alias{getData_editions}
\title{Harvest CBLOL editions data}
\usage{
getData_editions(ano = c(2014:2020))
}
\arguments{
\item{ano}{integer.}
}
\value{
A tibble containing: year,tournament(split/playoff), prize pool(R$), winner, runner up and league.
}
\description{
Creates a tibble containing Leaguepedia data on each CBLOL edition.
}
\examples{
a <- getData_editions()
b <- getData_editions(ano = c(2020,2018))
}
