% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_games.R
\name{getData_games}
\alias{getData_games}
\title{Harvest data of CBLOL games from Leaguepedia}
\usage{
getData_games(Team = NULL, Year, Split)
}
\arguments{
\item{Team}{(character) Its very case sensitive and the name(s) have to be passed exactly as it is written in Leaguepedia}

\item{Year}{(numeric) The year of the edition you want to access data (2015:2020).}

\item{Split}{(character) Use if you want to specify a Split: "Split 1", "Split 2", "Split 1 Playoffs","Split 2 Playoffs"}
}
\value{
A tibble containing: date, split, patch, blue team, red team, winner, bans blue, bans red, picks blue, picks red, blue roster, red roster and league.
}
\description{
Creates a tibble containing Leaguepedia data on CBLOL games.
}
\examples{
games <- getData_games(Year = 2020, Split = c("Split 2","Split 2 Playoffs"))
}
