% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_players.R
\name{getData_players}
\alias{getData_players}
\title{Harvest from Leaguepedia data of players in games of CBLOL}
\usage{
getData_players(Role, Year, Split, Playerid = NULL, Team = NULL)
}
\arguments{
\item{Role}{(character) The lane of the player. It should contain at least one of the five roles: "Top", "Jungle", "Mid", "Bot" and "Support".}

\item{Year}{(numeric) The year you want to access data(2015:2020).}

\item{Split}{(character) The split you want to access data: "Split 1", "Split 2", "Split 1 Playoffs" or "Split 2 Playoffs".}

\item{Playerid}{(character) Its very case sensitive and the playerid(s) have to be passed exactly as in Leaguepedia.}

\item{Team}{(character) Its very case sensitive and the name(s) have to be passed exactly as in Leaguepedia}
}
\value{
A tibble containing: player, team, role,  number of games played, victories, defeats, win rate, kills, deaths, assists, KDA, CS per game, CS per minute, gold per game, gold per minute, kill participation, percentage of kills/team, percentage of gold/team, number of champions played, year, split, league.
}
\description{
Creates a tibble containing Leaguepedia data of players in CBLOL games
}
\examples{
a <- getData_players()
b <- getData_players(Role = "Top",Year = 2020, Split = "Split 2")
c <- getData_players(Playerid = "Robo",Year = c(2019,2020))
d <- getData_players(Team = "INTZ",Year = 2020, Split = "Split 2")
}
