% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_rosters.R
\name{getData_rosters}
\alias{getData_rosters}
\title{Harvest data of CBLOL rosters from Leaguepedia}
\usage{
getData_rosters(
  Playerid = NULL,
  Team = NULL,
  Split = c("Split_1", "Split_2"),
  Year = c(2015:2020),
  Role = c("Top Laner", "Jungler", "Mid Laner", "Bot Laner", "Support", "Coach")
)
}
\arguments{
\item{Playerid}{(character)  Its very case sensitive and the playerid(s) have to be passed exactly as it is written at Leaguepedia}

\item{Team}{(character)  Its very case sensitive and the name(s) have to be passed exactly as it is written at Leaguepedia}

\item{Split}{(character)  Use if you want to specify a Split: "Split_1" or "Split_2".}

\item{Year}{(numeric) The year of the edition you want to access data}

\item{Role}{(character)  Use if you want to specify one or more roles: "Top Laner", "Jungler", "Mid Laner", "Bot Laner", "Support" or "Coach".}
}
\value{
A tibble containing: country, playerid (id), name, team, year, split, role and league. If there is no data for the entry you specified it will return a message saying "There is no data for this entry".
}
\description{
Creates a tibble containing Leaguepedia data on CBLOL rosters on each edition
}
\examples{
a <- getData_rosters()
b <- getData_rosters(Year = 2020,Split = "Split_1")
c <- getData_rosters(Playerid = "brTT")
d <- getData_rosters(Team = "INTZ",Role = "Coach")
}
