% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_titlesOrg.R
\name{getData_titlesOrg}
\alias{getData_titlesOrg}
\title{Harvest data of CBLOL wins per Organization from Leaguepedia}
\usage{
getData_titlesOrg(Team = NULL)
}
\arguments{
\item{Team}{(character) Its very case sensitive and the name(s) have to be passed exactly as it is written in https://lol.gamepedia.com/Circuit_Brazilian_League_of_Legends}
}
\value{
A tibble containing: team, x1st_place, x2nd_place, x3rd_4th_place, league
}
\description{
Creates a tibble containing Leaguepedia data on CBLOL wins per Organization
}
\examples{
titlesOrg <- getData_titlesOrg()
}
